Version final
Creo

+++ PBl2_9. taldea/PBl2_9. taldea/Erabilgarriak.h
@@ -14,16 +14,16 @@
 #define TERMINATE -1
 SDL_Renderer* renderer;
 
-#define MUGIT_IMAGE "../Media/Images/mugit.bmp"
-#define MENU_IMAGE "../Media/Images/Menu.bmp"
-#define FILE_IMAGE "../Media/Images/helbidea.bmp"
-#define MAP_IMAGE "../Media/Images/irudia.bmp"
-#define DIM_IMAGE "../Media/Images/MapaSortu.bmp"
+#define MUGIT_IMAGE "../Media/Images/mugit.bmp" // Robota
+#define MENU_IMAGE "../Media/Images/Menu.bmp" // Menu nagusiaren irudia
+#define FILE_IMAGE "../Media/Images/helbidea.bmp" // Fitxategia eskatzeko menuaren irudia
+#define MAP_IMAGE "../Media/Images/irudia.bmp" // Maparen irudia eskatzeko menuaren irudia
+#define DIM_IMAGE "../Media/Images/MapaSortu.bmp" // Dimentsioak eskatzeko menuaren irudia
 
 typedef struct DE_DIMENTSIOA
 {
-	long width;
-	long height;
+	int width;
+	int height;
 } DIM;
 
 typedef struct DE_POSIZIOA
@@ -46,10 +46,15 @@ typedef struct A_MUGIMENDUA
 	struct A_MUGIMENDUA* ptrHurrengoa;
 } MUGI, * ptrMugi;
 
+// ptrMugi motako aldagaiak
 void askatuMugitu(ptrMugi* burua);
+// ptrPuntua motako aldagaiak askatu
 void askatuPuntuak(ptrPuntua* burua);
+// Matriz baten memoria erreserba egin
 int erreserbaBurutu(float** bek, int kop);
+// Puntu bat bilatu eta haren koordenatuak jaso
 void bilatu(ptrPuntua burua, float* x, float* y, int id);
+// Irudi bat sortu
 int irudiaSortu(float x, float y, char* name, SDL_Window* window);
 
 #endif
